renamed app to api and ran tests succesffully
