+++ api/utils/mcq_generator.py
@@ -3,9 +3,9 @@
 import random
 from typing import List, Optional, Tuple
 
-from app.data.mcq_graph import MCQGraph
-from app.models import MCQ, MCQRelationship
-from app.utils.fake_word_generator import FakeWordGenerator
+from api.data.mcq_graph import MCQGraph
+from api.models import MCQ, MCQRelationship
+from api.utils.fake_word_generator import FakeWordGenerator
 
 
 class MCQGenerator:
