+++ frontend/templates/error.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>MCQBot</title>
+  </head>
+  <body>
+    <h1>There was an issue generating a multiple choice question. Please try refreshing the page. </h1>
+  </body>
+</html>
