+++ frontend/templates/home.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+    <style>
+        .btn {
+            display: block;
+            margin: 5px 0;
+            padding: 10px;
+            border-radius: 5px;
+            border: 1px solid #ccc;
+            width: 100%;
+            max-width: 300px;
+        }
+        .correct {
+            color: green;
+        }
+        .incorrect {
+            color: red;
+        }
+    </style>
+</head>
+<body>
+
+    <h1>Which of the following links to {{ topic }}?
+    </h1>
+
+    <form>
+        {% for item in choices %}
+            <button type="button" class="btn" id="{{ item }}" onclick="checkAnswer('{{ item }}')">{{ item }}</button>
+        {% endfor %}
+
+         <button id="refresh" class="btn" onclick="location.reload()">Refresh</button>
+    </form>
+       
+
+
+
+    <script type="text/javascript">
+        var correctAnswer = "{{ answer }}";
+        function checkAnswer(answer) {
+            if (answer == correctAnswer) {
+                $("#" + answer).addClass("correct");
+            } else {
+                $("#" + answer).addClass("incorrect");
+            }
+            $("button").attr("disabled", "disabled");
+            $("#" + correctAnswer).addClass("correct");
+            $("#refresh").prop("disabled", false);
+        }
+    </script>
+
+</body>
+</html>
