+++ tests/test_neo4j/neo4j_connect.py
@@ -1,6 +1,6 @@
 """Module containing utility functions of neo4j database connection"""
 # pylint: disable=broad-exception-caught broad-exception-raised
-from app.data.neo4j_graph import Neo4JGraph
+from api.data.neo4j_graph import Neo4JGraph
 from dotenv import load_dotenv
 
 load_dotenv()
